bloom

+++ pkg/bloom/bitset.go
@@ -8,58 +8,58 @@ const (
 // todo should we just wrap big.Int instead?
 
 type BitSet struct {
-	length uint
+	length uint64
 	set    []uint64
 }
 
-func NewBitSet(length uint) *BitSet {
+func NewBitSet(length uint64) *BitSet {
 	return &BitSet{
 		length: length,
 		set:    make([]uint64, wordsNeededFor(length)),
 	}
 }
 
-func wordsNeededFor(i uint) int {
+func wordsNeededFor(i uint64) int {
 	if i > maxCap() {
 		panic("bitset max cap exceeded")
 	}
 	return int((i + wordSize - 1) >> log2WordSize)
 }
 
-func indexInWord(i uint) uint {
+func indexInWord(i uint64) uint64 {
 	return i & (wordSize - 1)
 }
 
-func maxCap() uint {
-	return ^uint(0)
+func maxCap() uint64 {
+	return ^uint64(0)
 }
 
-func (b *BitSet) Set(i uint) {
+func (b *BitSet) Set(i uint64) {
 	if i >= b.length {
 		b.extendToFit(i)
 	}
 	b.set[i>>log2WordSize] |= 1 << indexInWord(i)
 }
 
-func (b *BitSet) Clear(i uint) {
+func (b *BitSet) Clear(i uint64) {
 	b.set[i>>log2WordSize] &^= 1 << indexInWord(i)
 }
 
-func (b *BitSet) SetTo(i uint, value bool) {
+func (b *BitSet) SetTo(i uint64, value bool) {
 	if value {
 		b.Set(i)
 	}
 	b.Clear(i)
 }
 
-func (b *BitSet) Get(i uint) bool {
+func (b *BitSet) Get(i uint64) bool {
 	if i > b.length {
 		return false
 	}
 	return b.set[i>>log2WordSize]&(1<<indexInWord(i)) != 0
 }
 
-func (b *BitSet) extendToFit(i uint) {
+func (b *BitSet) extendToFit(i uint64) {
 	if i >= maxCap() {
 		panic("bitset max cap exceeded")
 	}
